requirements.txt
openai==1.60.2
python-dotenv==1.0.1
pytest==7.4.3
pyyaml==6.0.1